Wire everything into the DefinitionService
